Initial changes to build bash as C++03 or C++11.
Changes to build bash as C++03 or C++11 instead of K&R C. The major
required change was converting K&R prototypes into C++ prototypes
and making all of the external prototype definitions compatible.
Other changes to fix build compiler errors/warnings include:

- generic list data type now uses C++ inheritance for type safety.
  But it doesn't seem to be used by anyone other than WORD_LIST.
- unwind protection now has three functions, depending on whether
  the unwind function takes no params, an int param, or a pointer.
- removed macros to handle pre-ANSI compilers that interfere with
  VS Code analysis, like PARAMS().
- removed "register" keyword from vars and reduced their scopes.
- updated y.tab.c and y.tab.h with GNU bison 3.8.2 from parse.y.
- added "const" to pointers where possible, especially strings.
- changed boolean variables and params from "int" to "bool" type.
- #ifdef'd out or deleted a few obsolete/useless hacks, such as the
  code in main() that opens and closes "/dev/tty", and then sets
  the no-close-on-exec flag on a range of fds for an old SunOS bug.

I'm testing on Ubuntu Linux 22.04 with GCC 11.2.0 and clang 12.0.1.
Configured with "--without-bash-malloc" to use the system malloc(),
and with "--with-included-gettext" to fix the libintl C++ issues.
I had to copy a newer libintl/plurals.y parser to regenerate the
parser sources using a current version of Bison.

Unfortunately this version immediately fails in the address sanitizer
for me, but I wanted to push my work so far to my git repo now that
it successfully compiles and links as C++03.

Testing with "-fstack-protector-strong -fsanitize=address" and with
both "-std=c++03" and "-std=c++11", on both C++ compilers, with and
without "--disable-job-control", for a total of 8 combinations.