Small refactorings, plus fix a bug I introduced.
I accidentally commented out some code in redir.c with important
side effects, because of how I'd refactored the return value.

Also updated the variable code to use a union type already defined
in variables.h, and patched general.h to work with Linux glibc's
definition of group_member() in C++, as well as VMS, which doesn't
have that function.