Upgrade autoconf scripts for C++ and future work.
- Upgrade the autoconf tests and scripts to GNU autoconf 2.71.
- Update Makefiles to use the C++ compiler everywhere.
- Upgrade the bundled version of lib/intl to gettext-0.19.
- Remove obsolete autoconf tests for pre-ANSI C compilers.
- Add new autoconf tests for group_member(), fork(), and vfork().
- Update code to properly use either bundled or libc group_member().
- Add "-x c++" to suppress Clang "treating 'c' input as 'c++'" warning.
- Add new `--enable-vfork-subshell` option to test OpenVMS vfork() code.
- Fix lib/malloc code to compile as C++.

Note that when using `-fsanitize=address` in CFLAGS, it's necessary
to set `ASAN_OPTIONS=detect_leaks=0` in the environment to ./configure
so that the `GETCWD_BROKEN` and `HAVE_ICONV` tests don't fail due to
memory leaks in glibc on exit. Also, bash malloc immediately fails when
the address sanitizer is enabled (I've been using `--without-bash-malloc`).