Rename the builtins files from .def to _def.cc.
I've decided to rename the "builtins/*.def" files to end in "_def.cc"
so that syntax highlighting works properly, and to save an extra step
of generating the .cc file when it can be compiled directly instead.

For this to work, I've prepended "// " to the beginning of directive
lines that "mkbuiltins" uses to generate "builtext.h" and the embedded
docs files. I'm currently working on updating mkbuiltins to strip the
the C++ comment lines from the beginning of the newly-renamed files.