Convert Automake config to use subdir-objects.
Replace the per-directory Makefile.am files with a single
top-level Makefile.am. This required a bit of trial and error
to get the directory paths and dependencies correct, especially
the proper usage of $(CONFIG_HEADER) in the dependencies for
the files generated by mkbuiltins and psize, which require the
"config.hh" header file to be generated before anything else.

I also added some entries to .gitignore to ignore strange .deps
directories and .dirstamp files that the build is now putting
in the source directory tree (when building from a separate objdir).