Updates for MINIX compatibility.
MINIX 3.4.0 needs <cstdlib> included for std::exit() (psize.cc),
and it fails the WCWIDTH_BROKEN configure test, which I'm going
to fix shortly (it needs to reference the Readline object).

I also lowered the version requirement of m4/threadlib.m4 from
Autoconf 2.71 to 2.69, although I'm not sure it actually works
with that version, given all the features I'm using. At least I'm
able to build on MINIX with Autoconf 2.69 and aclocal-1.15 installed,
as long as I don't try to regenerate any of the config scripts.