Fix Makefile.am to put .o files in the right place.
Remove the $(srcdir) prefix from the source filenames.
This puts the .o and other build files in the correct place
when building out-of-tree.

I'm working on fixing mkbuiltins.cc to use the passed -D argument
to find the builtins files, now that it's getting the list relative
to the source directory. I'll prepend the srcdir path from -D.