Further cleanups to build config scripts and code.
Add back tests for functions and header files that we actually use.
Regenerate Makefile.in, configure, and config.hh.in with autoreconf.
Fixed various compiler warnings.

Added manual dependencies for files that include "bashhist.hh", etc.
because the BUILT_SOURCES target isn't reliable when the build tools
that generate the .hh files require config.hh in order to build.
With this change, parallel builds appear to work well.

I've also removed code for AFS and Linux <libaudit.h> support, neither
of which seem particularly useful. If you want to audit bash, there
are other ways to do it than a code path that's not built by default.