Move custom parser code from parse.yy into parser.cc.
Create a new parser.cc file (to go with the existing parser.hh)
that holds the custom tail end of parse.yy, so that parse.yy contains
only the portion that needs to be run through Bison. This makes it
easier to work on the custom bash parser code directly as a .cc file.