Big changes to make the generated parser compile.
I've updated parse.yy to be fully C++ type-compatible, without having
to make too many changes. I'm hoping that any future changes won't lead
to major changes in the line numbering or code of the checked-in
generated parse.cc and parse.hh files.

The biggest change in terms of files touched was converting all uses
of intmax_t (and uintmax_t) to int64_t / uint64_t. This is a good idea
both in terms of giving all versions of bash the same behavior, but
also because glibc is moving to a 256-bit type for intmax_t, which is
definitely not what anyone wants here. We really want exactly 64 bits.