Change WORD_LIST, PATTERN_LIST, and REDIRECT back to GENERIC_LIST.
Revert an ill-advised change I started making to try to use std::vector
instead of singly-linked lists that the code often has to reverse. Now
WORD_LIST, PATTERN_LIST, and REDIRECT are children of GENERIC_LIST,
with their own non-virtual methods to reverse the list (returning the
new list head), append an element to the end, and in the case of
WORD_LIST, a method to prepend a newly-created WORD_LIST object
encapsulating the specified WORD_DESC (and return the new object).