Minor fixes to bring the code closer to the original.
Add a "noexcept" to one of the new class destructors, and remove
the "prepend()" method in favor of a new WORD_LIST constructor
that takes a WORD_DESC* and a WORD_LIST* like make_word_list() did.

I'm not sure why the Makefile decided to change paths for the #line
pragmas. It used to put files in my objdir, and now it's directly
updating the parse.cc and parse.hh in my srcdir. I used to have to
move the updated versions into the srcdir. Strange.