Reformat all C++ source files to GNU coding style.
Reformat all C++ source files using clang-format from LLVM 14.0.0.
This is a large diff, but I wanted to make it sooner, rather than
later, considering that the programming language has changed from
C to C++, and I wasn't sure what formatting to use for things like
exceptions and namespaces.

I had to manually fix some of the builtins/*_def.cc files to not
word wrap the $SHORT_DOC lines, but otherwise this is the result
of clang-format-14 plus some build failure fixes due to missing
definitions after clang-format rearranged the header files.

The main cause of this diff's size appears to be the replacement
of tab characters with spaces, which are much easier to deal with.

I've also made some minor changes to the Shell class with this diff,
which I'm going to continue working on, and fixed mkbuiltins to
emit builtins.cc that includes "bashintl.hh", not "bashintl.h".