C++ and Infer fixes for shell.cc and readline.
Finished refactoring for shell.cc to compile. Updated lib/readline
so that it always implements the virtual functions in ReadlineShell,
so that shell.cc can dynamic_cast itself to a Readline *. I also
removed a check for HAVE_FCNTL, which I removed from the configure
script, because bash and Readline require a working fcntl().

I removed several dead stores identified by Infer, and added comments
to indicate a dead store and a file open resource leak that I believe
are false alarms. I may try to tighten up BUFFERED_INPUT so that it's
guaranteed to close the file descriptor it owns on destruction.