Fix more C++ build failures, up to general.cc.
I'm trying to fix the source files in their order in Makefile.ac,
with everything up to general.cc now compiling. parser.cc is the
source file that I'm currently working on, which contains the
C code from parse.y that doesn't need to be processed by Bison.

I'm slightly concerned about the special alrm_catcher() code in
eval.cc that can longjmp to the top level to terminate directly
from the signal handler itself, which doesn't seem safe. This
code path also appears to only be used when the "TMOUT" variable
is set, so it may not be very well tested to begin with.

I've changed the code to set the volatile sig_atomic_t to make it
appear that a SIGKILL has been received, so that hopefully the
next call to test those variables will throw a program-exiting
exception that eval.cc can catch for the shell to exit correctly.
It's not possible for me to throw an exception or otherwise make
a longjmp() call directly from the POSIX signal handler without a
place to jump to, and I'd prefer to remove setjmp/longjmp entirely.

I've updated config.h and other .h filenames to their new names
in various source files that don't build yet, e.g. loadables.