Updates for C++03 and continued parser refactoring.
I had to add a new "CONSTEXPR" that gets replaced with either
"constexpr" (C++11) or nothing (C++03), to handle statically-initialized
arrays that I'd like to define as constexpr. In every other place in the
code where I'm using constexpr, it makes more sense to replace it with
"const" for C++03, but that creates a syntax error here, so I added a
new all-caps macro as a hack for defining static arrays.

More updates to parser.cc, after which I'm planning to fix the small
files in mostly alphabetical order. My plan for the parsing code is to
pass and return std::string wherever possible, and to replace illegal
"magic pointer" error return values that signify errors with custom C++
exceptions for each error type. If any slip through to the top-level
catch handler, it'll print the name of the uncaught exception.