Continue lib/readline work and switch to size_t.
Fixes to lib/readline to update the callback API that it provides
to bash (and other users). Changed from "unsigned int" to size_t
for clarity and to avoid any overhead casting between 32 and 64 bits.
This also fixes build failures on stricter compilers that don't
implicitly cast unsigned int to size_t, e.g. in string append ().

The other major change was to use std::string instead of char *, except
when it required major rewrites that I didn't want to tackle right now.

I also changed "sighandler_t" to "SigHandler" in a few cases, because
Solaris and other systems don't have a typedef for sighandler_t.