Add includes for std::string_view and string-view-lite.
Update the makefiles to add a new include/nonstd/string_view.hpp
file, which is now included for pre-C++17 compilers as a substitute
for <string_view>, which I'm going to be using in upcoming cleanups.
I've modified the .hpp file to define the "sv" operator in std::,
so that it should be compatible with the std:: version for defining
string constants.

Copied the .hpp from: https://github.com/martinmoene/string-view-lite
It's under a Boost license, which is GPL-compatible.