Checkpoint execute_cmd.cc refactoring.
Continued refactoring the COMMAND hierarchy and execute_cmd.cc.
I've added a type enum back to the parent COMMAND class so that
the current C code that uses the type in switch statements can be used
in the C++ version with no code changes and the same performance.

I was trying to use RTTI (typeid and dynamic_cast<>) without the type
enum, but that approach didn't seem to offer any benefit, so I changed
back to the C version's enum, with static_cast<> to subclasses after
checking the type.