Refactor evalfile and evalstring to use exceptions.
Refactor the logic for evalfile and evalstring functions to free
allocated memory and restore the previous state on error using C++
exceptions instead of setjmp/longjmp.

Switched back to using C strings where it's easier, including for the
string to parse (if the SEVAL_NOFREE flag is set, it isn't deleted).

I still need to finish C++ifying the code in the EvalFileUnwindHandler
in builtins/evalfile.cc, but it now compiles, with some #if 0 parts.