Add nstrftime and remove strpbrk modules.
The strpbrk gnulib module is marked obsolete because every C/C++ compiler
should have the function. I've added "nstrftime" and changed the code
from using strftime () to nstrftime (), with the extra parameters set
to nullptr and 0, so that it will give the same results everywhere,
and I won't have to worry about buggy OS strftime () implementations.