Update headers for Haiku OS compatibility.
Haiku OS defines "B_ERROR" as a macro, so I renamed the bstream_flags
enum values to start with "BST_" instead of "B_", which should be
slightly more memorable as well as not conflicting with BeOS errors.

I also added #include <fcntl.h> to a common bash header file so that
mkbuiltins.cc can find open() on Solaris 10.